+++ measureme/src/profiler.rs
@@ -83,7 +83,7 @@ impl<S: SerializationSink> Profiler<S> {
 
     /// Records an event with the given parameters. The event time is computed
     /// automatically.
-    pub fn record_instant_event(&self, event_kind: StringId, event_id: StringId, thread_id: u64) {
+    pub fn record_instant_event(&self, event_kind: StringId, event_id: StringId, thread_id: u32) {
         let raw_event =
             RawEvent::new_instant(event_kind, event_id, thread_id, self.nanos_since_start());
 
@@ -96,7 +96,7 @@ impl<S: SerializationSink> Profiler<S> {
         &'a self,
         event_kind: StringId,
         event_id: StringId,
-        thread_id: u64,
+        thread_id: u32,
     ) -> TimingGuard<'a, S> {
         TimingGuard {
             profiler: self,
@@ -136,7 +136,7 @@ pub struct TimingGuard<'a, S: SerializationSink> {
     profiler: &'a Profiler<S>,
     event_id: StringId,
     event_kind: StringId,
-    thread_id: u64,
+    thread_id: u32,
     start_ns: u64,
 }
